refactoring tests for the redispatch
