+++ grid2op/tests/__init__.py
@@ -1,3 +1,10 @@
-__all__ = ["BaseBackendTest"]
+# Copyright (c) 2019-2020, RTE (https://www.rte-france.com)
+# See AUTHORS.txt
+# This Source Code Form is subject to the terms of the Mozilla Public License, version 2.0.
+# If a copy of the Mozilla Public License, version 2.0 was not distributed with this file,
+# you can obtain one at http://mozilla.org/MPL/2.0/.
+# SPDX-License-Identifier: MPL-2.0
+# This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
-# import grid2op.tests.BaseBackendTest
+__all__ = ["BaseBackendTest",
+           "BaseRedispTest"]
